Cuckoo barrage!
