+++ cuckoobird.cabal
@@ -32,7 +32,8 @@ executable cuckoobird
 
     -- Modules included in this executable, other than Main.
     other-modules:
-      CuckooLib
+        CuckooLib
+      , CuckooNest
 
     -- LANGUAGE extensions used by modules in this package.
     other-extensions: OverloadedStrings
